Changed parser/converter API to work without a document instance
Parsers and converters can now work without the Kramdown::Document class
which is just provided for convenience. All the needed information is
stored in the element tree itself. Information that has no direct
representation as an element is stored in the options of the root
element (e.g. footnote definitions).

The main interface to kramdown, the Kramdown::Document class, was
internally changed but still has the same API. This means that
applications that only used this interface will still work after
updating.