Backwards incompatible change: code language now specified via special class
Using lang='LANGUAGE' for specifying the code language of code
blocks/spans is not good because it conflicts with the normal
usage of the lang attr.

Therefore the code language must now be specified via a special
class 'language-LANGUAGE' which is also the way how it is done
in the HTML5 spec.