Fixed GH#91: template extension should be the same as converter name
When specifying the name of ConverterName on the CLI, it is needed to
use converter_name. However, the used template extension was
convertername. This is fixed now so that both versions work (for
backwards-compatibility).

The former way is deprecated and support for templates with such an
extension will be removed in 2.0.