Adjusted reference link parsing to work more similar to inline links
Single and double quotes are now allowed in both forms, provided
they are not preceeded by a space (they were already allowed in
inline links although this was not documented). The syntax
document has been adjusted accordingly.

Additional, the syntax document has been adjusted for reference
links to mention that the spaces/tabs between link ID in brackets
and the URL are optional.

Fixes #239