Update header ID generation for GFM parser
GFM uses a completely different ID generation model than kramdown. To
make ID generation more compatible, the GFM parser is adapted using the
algorithm that Github uses.

This also means that the option 'auto_id_stripping' is always set to
true for the GFM parser.

There will also be problems with links of existing documents that use
the GFM parser.