New math-to-HTML/CSS/MathML engine: SsKaTeX
The name SsKaTeX means server-side KaTeX, as opposed to client-side KaTeX.
It eliminates the need (and the flexibility, and the overhead)
for client-side math-rendering Javascript.

Consider this a lightweight alternative to Mathjax-Node.
SsKaTeX uses KaTeX's katex.min.js (not included) instead of MathJax
and uses a configurable or auto-selected JS engine (via ExecJS),
e.g. Duktape, rather than depending strictly on Node.js.

Javascript execution context initialization is done only once.
With some JS engines, no inter-process communication is needed
because the Javascript is handled by library functions.
As a result, the performance is reasonable.