GFM parser: quirk to disable auto typographic symbol conversion
With GFM, we want to be as close as possible to GitHub Flavored
Markdown. In particular, GitHub does not replace -- with en-dash,
--- with em-dash et cetera, and doing this not only breaks
HTML rendering (resulting in, say, –option instead of --option).

The above replacement also makes header IDs incompatible with those
generated by GitHub, making it way harder to create a document
with intralinks that work both on GitHub and kramdown-generated
HTML.

So, let's introduce another GFM quirk to turn off automatic
typographic symbol conversion. NOTE it is off by default, i.e.
needs to be explicitly enabled.

For more details on the issue and examples, see
https://github.com/gettalong/kramdown/issues/459

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>