New option :auto_dd_para to get automatic :p in :dd elements
Without the new option, as is the default, the first paragraph in
a multi-paragraph dd entry is not marked as a :p block unless
explicitly requested with an empty line before the dd entry.

With the new option, having more than one paragraph in the dd
entry causes the first paragraph to be p-marked as well.
The explicit syntax with a preceding empty line remains available,
but is only necessary for single-paragraph items with this option.