Don't require whitespace before closing ATX ### marks
Prior to 135f1fd, ATX headers with closing ### marks did not require
preceding whitespace. While this is a requirement in GFM markdown[1],
the original markdown spec[2] doesn't require it.

This change therefore adds test cases to cover the reinstatement of this
behaviour, and adjusts the trimming regex to exclude closing ### marks
only if they are preceded by a backslash (for which a test case already
exists).

[1] https://github.github.com/gfm/#atx-heading
[2] https://daringfireball.net/projects/markdown/syntax