Handle the Kramdown::Document.new input option like other options
I was trying to understand how to switch parsers, but I was confused when I
couldn't find `input` in the Options documentation:
https://kramdown.gettalong.org/options.html I was instead able to find it in
the kramdown-gfm documentation: https://github.com/kramdown/parser-gfm but I
still wanted to know what the default value was - and to check that it was a
supported part of the API.

It was only after looking at the source code for a while that I realised
it is documented.

Aside from helping people with the problems I encountered above, this means
this argument is parsed and validated in a more standard way.