Add new option, `footnote_link_text`
`footnote_link_text` is a string that is directly inserted in
front of the footnote number. It can also be a formatted string
that the number is formatted into.

For example, if the string is "[footnote %s]", the footnote link
will be "[footnote 1]".

It will determine it the option is a forat string by running it
through `sprintf`; if the output is the same, then it's not a
format string.

Also adds tests to make sure of the new output.